Convert everything to BigDecimal

+++ db/migrate/20170809210510_create_orbits.rb
@@ -2,8 +2,8 @@ class CreateOrbits < ActiveRecord::Migration[5.1]
   def change
     create_table :orbits do |t|
       t.references :orbitable, polymorphic: true, index: true
-      t.integer :mass
-      t.integer :radius
+      t.decimal :mass
+      t.decimal :radius
 
       t.timestamps
     end
